Added JS and sample code for word cloud
Current plan is to serve the page then fetch the top words in java
script, then populate the cloud - and then do this on a timer so the
page auto-updates. This means the top words are available for anyone to
use & parse on /wc (at the moment).

It may actually be better to use a template to populate the available
words server-side before serving the page - less requests, no public
facing API.