Attempt to fix the memory leak
It isn't fixed, tho, cos Python memory fragmentation. Trying to limit
it by reusing the dictionary of words and just resetting the count of
each word to 0, rather than creating a new one.

Moved the Timer to the end of the `update_data` method. Prevents the
method queuing up (say, if we have to save data for the day - this
could take a few seconds).