Now using a Cron job to poll ADN every minute
This seems to be effective, but it has some problems. We should
definitely be saving to datastore, so that if the instance goes down we
can carry on. Secondly, I don't think there's any guarantee of the
instance staying up just because this script is polling it every
minute. Thirdly, if there's > 200 posts since the last time we queried
ADN, we're going to miss some posts.